Changed MonthlyPrice to decimal

+++ src/backend/SportLink.Backend/SportLink.API/Data/Entities/TrainingGroup.cs
@@ -11,7 +11,7 @@ public class TrainingGroup : BaseEntity
     public int AgeFrom { get; set; }
     public int AgeTo { get; set; }
     public SexEnum Sex { get; set; }
-    public int MonthlyPrice { get; set; }
+    public decimal MonthlyPrice { get; set; }
     public string Description { get; set; }
     public int OrganizationId { get; set; }
     public int SportId { get; set; }
